Added double type

+++ parser/src/ast_utils.rs
@@ -26,7 +26,7 @@ pub fn build_args(scanner : &mut Lex, stmt : &mut AstStmt, end : Token, syntax :
             },
             
             Token::FloatL(val) => {
-                let arg = ast::create_float(val as f32);
+                let arg = ast::create_float(val);
                 
                 if in_array {
                     current_arg.sub_args.push(arg);
